move utils/logger.js to src/lib/logger.js, update code and docs
